requirements.txt
Appium-Python-Client~=2.8.1
selenium~=4.8.2
pytest~=7.2.2